fix(sidebar): reconcile the status hide-list with the live statuses
Review follow-up:

- Read the hide-list through one reconciliation helper: ids of deleted
  statuses are dropped, and a list that would hide every surviving
  status is treated as no filter. The click-time "keep one visible"
  guard could not see a later status deletion coming, which left the
  sidebar empty with every checkbox unchecked.
- The filter badge counts reconciled ids, so a deleted status no longer
  leaves a phantom filter with no row to untick.
- Apply the filter in the workspace board and the Cmd+J palette, which
  build their own lists and documented that they mirror the sidebar.
- Keep the active workspace's row visible: "Move to status" sits on the
  row itself, so marking the workspace you are in as a hidden status
  deleted its row while its panes stayed open.
- Disable the last visible status checkbox instead of swallowing the
  click, and return the same array on a rejected toggle so it no longer
  churns a persisted-UI write.
